backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.14
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.2
bcrypt==4.1.1
PyJWT==2.11.0
psycopg2-binary==2.9.9
